allow repo_type of git@github.com
