+++ src/lean_dojo/utils.py
@@ -16,7 +16,6 @@ from functools import cache
 from contextlib import contextmanager
 from ray.util.actor_pool import ActorPool
 from typing import Tuple, Union, List, Generator, Optional
-from urllib.parse import urlparse
 
 from .constants import NUM_WORKERS, TMP_DIR, LEAN4_PACKAGES_DIR, LEAN4_BUILD_DIR
 
@@ -143,33 +142,6 @@ _CAMEL_CASE_REGEX = re.compile(r"(_|-)+")
 def camel_case(s: str) -> str:
     """Convert the string ``s`` to camel case."""
     return _CAMEL_CASE_REGEX.sub(" ", s).title().replace(" ", "")
-
-
-def repo_type_of_url(url: str) -> Union[str, None]:
-    """Get the type of the repository.
-
-    Args:
-        url (str): The URL of the repository.
-    Returns:
-        str: The type of the repository.
-    """
-    parsed_url = urlparse(url)
-    if parsed_url.scheme in ["http", "https"]:
-        # case 1 - GitHub URL
-        if "github.com" in url:
-            if not url.startswith("https://"):
-                logger.warning(f"{url} should start with https://")
-                return
-            else:
-                return "github"
-        # case 2 - remote URL
-        elif url_exists(url):  # not check whether it is a git URL
-            return "remote"
-    # case 3 - local path
-    elif is_git_repo(Path(parsed_url.path)):
-        return "local"
-    logger.warning(f"{url} is not a valid URL")
-    return None
 
 
 @cache
